stubbing the client library

+++ base/log.h
@@ -51,7 +51,7 @@ namespace eqBase
 }
 
 #define LOG_EXTRA  << getpid()  << "." << pthread_self() << " " << __FILE__ \
-  << " " << __LINE__ << ": "
+                   << " " << (int)__LINE__ << ": "
 
 #define ERROR (eqBase::Log::level >= eqBase::LOG_ERROR) && \
     std::cout << "[E]" LOG_EXTRA
